feat: fix minor errors

+++ Responsavel/Responsavel.java
@@ -21,7 +21,7 @@ public abstract class Responsavel
         return this.nome;
     }
 
-    public abstract double calcularDesconto(Imovel imovel, int idade1, int idade2, float porcentagemDesconto1, float porcentagemDesconto2);
+    public abstract double calcularDesconto(Imovel imovel, int idade1, int idade2, double porcentagemDesconto1, double porcentagemDesconto2);
 
     public abstract String toString();
 }
